Add files via upload
In this work, the  implementation of a a 2:1 multiplexer using a simplified neural network structure and a basic threshold function. The goal was to showcase how a neural network-like approach could be employed to model the behavior of a 2:1 multiplexer.

1) Activation Function
The work  uses  a simple threshold function as the activation function. The threshold function determines the output based on whether the weighted sum of inputs and a bias is above a certain threshold.

2) Neural Network Parameters:
The neural network had two inputs (input1 and input2) and a selector (selector). Weights (w1, w2, wsel) and a bias (bias) were introduced to control the behavior of the neural network.

3)Weighted Sum Calculation:
The weighted sum was calculated as sum = (input1 * w1) + (input2 * w2) + (selector * wsel) + bias.

4)Activation Function Application:
The result of the weighted sum was passed through the threshold function using neural_network function, which returned either 0 or 1 based on the activation threshold.

5)Verification:
The correctness of the implementation was verified using the CBMC solver and assertions. The expected behavior of the neural network for different input combinations was asserted to ensure proper functioning.